requirements.txt
python-docx>=1.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
requests>=2.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
docx2pdf>=0.1.8
sqlalchemy>=2.0.0
